+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Install iproute2 (`ip`) in the `cli` profile.
 - Install ping in the `cli` profile.
 - Install rsync in the `cli` profile.
+- Install bzip2 in the `cli` profile.
+- Install XZ Utils (`xz`) in the `cli` profile.
 
 ## [1.6.0] - 2026-08-10